spruce up test utils file organization
